feat: Add IRC connection state enum
Introduces `IrcConnectionState` enum in [`include/irc.h`](include/irc.h:21) to represent the various states of an IRC connection. This includes states for disconnected, connecting, connected, registering, and registered.

This enum will be used to manage and track the lifecycle of the IRC client's connection, improving clarity and control over connection status.