feat(irc): Display user join messages in channel buffer
When a user joins an IRC channel, a message of the form "user has joined #channel" is now displayed in the corresponding channel buffer.

This provides immediate feedback to the user, improving awareness of channel activity. The change is implemented within the [`irc_process_buffer()`](src/irc.c:338) function by appending the join message and triggering a UI refresh if the event occurs in the active buffer.